oki
Nuevos Cambios

+++ DataEntitysDescuentos/VIEW_EXCEL_SAP.cs
@@ -15,7 +15,7 @@ namespace DataEntitysAplicaciones.DataEntitysDescuentos
         public int? vacio1 { get; set; }
         public int? vacio2 { get; set; }
         public int? vacio3 { get; set; }
-        public byte PORCENDESC { get; set; }
+        public double PORCENDESC { get; set; }
         public string simbolo { get; set; }
         public string FECINI { get; set; }
         public string FECFIN { get; set; }
