+++ DataEntitysDescuentos/VIEW_VALORPORCLIENTE.cs
@@ -7,7 +7,7 @@ namespace DataEntitysAplicaciones.DataEntitysDescuentos
         [Key]
         public long ID_MDDESCUENTO { get; set; }
         public string Material { get; set; }
-        public byte PORCENDESC { get; set; }
+        public double PORCENDESC { get; set; }
         public int CANT { get; set; }
         public double subtotal { get; set; }
         public double Descuento { get; set; }
